pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic>=2.11.4",
    "pymongo>=4.12.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
